requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-rest-framework==0.1.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
kombu==5.5.3
packaging==25.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
